requirements.txt
aiogram==2.25.2
python-dotenv
openai==0.28
g4f==0.1.9.1
sqlalchemy==2.0.29
aiosqlite==0.20.0